Change the way purchase history and help pane close
Now they have an X icon instead of "Tillbaka till matbutik" and don't send the user back home but instead just closes the pane and resumes where they were.

They also have a slight transparent background to indicate that the panes are just layered on top of the page they were at and isn't a whole different place in the app.